Add Sku (Stock Keeping Unit)

+++ Models/SpuAddDto.cs
@@ -7,9 +7,9 @@ namespace init_api.Models
         public string Name{ get; set; }
         [Required]
         [Range(0,1)]
-        public int Saleable { get; set; }
+        public byte Saleable { get; set; }
         [Range(0,1)]
-        public int Valid { get; set; }
+        public byte Valid { get; set; }
         public DateTime CreateTime {get;set;}
         public DateTime LastUpdateTime {get;set;}
         public SpuAddDto(){
