+++ Models/SpuDto.cs
@@ -6,8 +6,8 @@ namespace init_api.Models
         public Guid UUID {get;set;}
         [Required]
         public string Name {get;set;}
-        public int Saleable {get;set;}
-        public int Valid {get;set;}
+        public byte Saleable {get;set;}
+        public byte Valid {get;set;}
         public DateTime CreateTime {get;set;}
         public DateTime LastUpdateTime {get;set;}
     }
