+++ Models/SpuUpdateDto.cs
@@ -6,9 +6,9 @@ namespace init_api.Models
         public string? Name{ get; set; }
         [Required]
         [Range(0,1)]
-        public int Saleable { get; set; }
+        public byte Saleable { get; set; }
         [Range(0,1)]
-        public int Valid { get; set; }
+        public byte Valid { get; set; }
         public DateTime LastUpdateTime {get;set;}
         public SpuUpdateDto(){
             this.LastUpdateTime=DateTime.UtcNow;
